+++ package.json
@@ -26,7 +26,7 @@
     "debug": "babel-node ./src/index.js --nolazy --inspect-brk=9229",
     "pretest": "rm -f ./database/*.test.*; cross-env NODE_ENV=test ./node_modules/.bin/sequelize db:migrate; cross-env NODE_ENV=test node_modules/.bin/sequelize db:seed:all",
     "test": "cross-env NODE_ENV=test mocha test --exit",
-    "postest":"npm run clean:build",
+    "postest": "npm run clean:build",
     "cover": "cross-env NODE_ENV=test nyc mocha test --exit",
     "coverreport": "nyc report --reporter=lcov",
     "precoveralls": "mkdir coverage",
@@ -50,7 +50,7 @@
       "docs",
       "./src/domain",
       "./src/config",
-      "./src/http/middlewares/security.js"
+      "./src/http/middlewares/jwt.js"
     ]
   },
   "dependencies": {
@@ -59,9 +59,11 @@
     "bcrypt": "^2.0.0",
     "body-parser": "^1.18.2",
     "consign": "^0.1.6",
+    "cors": "^2.8.4",
     "crypto-js": "^3.1.9-1",
     "express": "^4.16.3",
     "express-winston": "^2.4.0",
+    "helmet": "^3.12.0",
     "moment": "^2.21.0",
     "passport": "^0.4.0",
     "passport-hash": "^0.6.1",
