+++ src/http/middlewares/httpSecurity.js
@@ -1,0 +1,7 @@
+import Cors from 'cors'
+import Helmet from 'helmet'
+
+module.exports = app => {
+  app.use(Cors())
+  app.use(Helmet())
+}
